Properly add the benchmark. Do some cleanup
