+++ test/server_sent_event/splitting_test.exs
@@ -1,19 +1,31 @@
 defmodule ServerSentEvent.SplittingTest do
-  alias ServerSentEvent.Splitting
-
   use ExUnit.Case, async: true
   use ExUnitProperties
 
   @moduletag :property
-  @moduletag :splitting
-  @moduletag :benchmark
 
   test "all approaches to splitting work the same way" do
-    check all input <- StreamData.string(:ascii),
+    check all input <- StreamData.string([?a..?f, ?ż, ?\r, ?\n]),
               max_runs: 500,
-              max_run_time: 600 do
-      assert Splitting.with_binary_split(input) == Splitting.with_regex(input)
-      assert Splitting.with_elixir_string_split(input) == Splitting.with_regex(input)
+              max_run_time: 100 do
+      assert with_binary_split(input) == with_regex(input)
+      assert with_elixir_string_split(input) == with_regex(input)
     end
+  end
+
+  @new_line_regex ~r/\R/
+  @new_line_sequences ["\r\n", "\r", "\n"]
+
+  def with_regex(stream) do
+    String.split(stream, @new_line_regex, parts: 2)
+  end
+
+  def with_binary_split(stream) do
+    # not global means just two parts
+    :binary.split(stream, @new_line_sequences)
+  end
+
+  def with_elixir_string_split(stream) do
+    String.split(stream, @new_line_sequences, parts: 2)
   end
 end
